Preparation for the Qbit base board communication
Modify the qbitInit() prototype by adding poiners to MicroBitMessageBus,
MicroBitSerial and MicroBitIO objects. A pointers to the object-members
from the global uBit object must be passed there.

The comminucation to the Qbit base-board is over the serial interface,
i.e. motor control and IR remote. The IO pins are for the ulrasonic
sensors and the extension pins on the top-board.